final commit

+++ MonitoringStationAPI/Models/AirPollutionData.cs
@@ -8,7 +8,7 @@ namespace MonitoringStationAPI.Models
         public int Id { get; set; }
         public int SensorId { get; set; }
         public DateTime Timestamp { get; set; }
-        public double AirPollution { get; set; }
+        public float AirPollution { get; set; }
         public bool TriggerThresholdWarning { get; set; }
 
     }
