+++ MonitoringStationAPI/Models/CO2EmissionsData.cs
@@ -8,7 +8,7 @@ namespace MonitoringStationAPI.Models
         public int Id { get; set; }
         public int SensorId { get; set; }
         public DateTime Timestamp { get; set; }
-        public double CO2Emissions { get; set; }
+        public float CO2Emissions { get; set; }
         public bool TriggerThresholdWarning { get; set; }
 
     }
