+++ MonitoringStationAPI/Models/HumidityData.cs
@@ -8,7 +8,7 @@ namespace MonitoringStationAPI.Models
         public int Id { get; set; }
         public int SensorId { get; set; }
         public DateTime Timestamp { get; set; }
-        public double Humidity { get; set; }
+        public float Humidity { get; set; }
         public bool TriggerThresholdWarning { get; set; }
 
     }
